collect file per obj
