+++ modules/scaleio/collect_sdc.go
@@ -1,0 +1,21 @@
+package scaleio
+
+var (
+	selectedSdcStatsQuery = `
+{
+    "selectedStatisticsList":[
+        {
+            "type":"Sdc",
+            "allIds":[
+            ],
+            "properties":[
+                "numOfMappedVolumes",
+                "userDataReadBwc",
+                "userDataWriteBwc",
+                "volumeIds"
+            ]
+        }
+    ]
+}
+`
+)
