+++ modules/scaleio/collect_storage_pool.go
@@ -1,0 +1,1 @@
+package scaleio
